pyproject.toml
[project]
name = "data-clean-env"
version = "1.0.0"
description = "Data Cleaning & Transformation OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "openai>=1.3.0",
    "python-multipart>=0.0.6",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["data_clean_env*", "server*"]
